Added a boot logo option
Must use a binary (.bin) image renamed to bootlogo.tns. This is NOT my creation I have just simply Proposed it as a change because it may be useful to some people. My credits are to ZephRay and her article on cncalc (http://www.cncalc.org/forum.php?mod=viewthread&tid=9774). There it illustrates how to create the binary file from a regular image.